+++ Source/AllOverIt.Serilog/Sinks/CircularBuffer/CircularBufferSink.cs
@@ -1,0 +1,43 @@
+﻿using AllOverIt.Assertion;
+using Serilog.Core;
+using Serilog.Events;
+using Serilog.Formatting;
+
+namespace AllOverIt.Serilog.Sinks.CircularBuffer
+{
+    /// <summary>A destination for log events that are stored in a fixed-sized circular buffer.</summary>
+    public class CircularBufferSink : ILogEventSink
+    {
+        private readonly ICircularBufferSinkMessages _sinkMessages;
+        private readonly ITextFormatter _formatter;
+
+        /// <summary>Constructor.</summary>
+        /// <param name="sinkMessages">The circular buffer to receive the log event messages.</param>
+        /// <param name="formatter">A custom formatter to apply to the log events.</param>
+        public CircularBufferSink(ICircularBufferSinkMessages sinkMessages, ITextFormatter formatter)
+        {
+            _sinkMessages = sinkMessages.WhenNotNull(nameof(sinkMessages));
+            _formatter = formatter.WhenNotNull(nameof(formatter));
+        }
+
+        /// <inheritdoc/>
+        public void Emit(LogEvent logEvent)
+        {
+            _ = logEvent.WhenNotNull(nameof(logEvent));
+
+            using (var writer = new StringWriter())
+            {
+                _formatter.Format(logEvent, writer);
+
+                var formattedMessage = writer.ToString().Trim();
+
+                var bufferMessage = new CircularBufferSinkMessage(logEvent, formattedMessage);
+
+                lock (_sinkMessages)
+                {
+                    _sinkMessages.PushBack(bufferMessage);
+                }
+            }
+        }
+    }
+}