+++ Source/AllOverIt.Serilog/Sinks/CircularBuffer/CircularBufferSinkMessages.cs
@@ -1,0 +1,12 @@
+﻿using AllOverIt.Collections;
+
+namespace AllOverIt.Serilog.Sinks.CircularBuffer
+{
+    internal class CircularBufferSinkMessages : CircularBuffer<CircularBufferSinkMessage>, ICircularBufferSinkMessages
+    {
+        public CircularBufferSinkMessages(int capacity)
+            : base(capacity)
+        {
+        }
+    }
+}