+++ Source/AllOverIt.Serilog/Sinks/CircularBuffer/ICircularBufferSinkMessages.cs
@@ -1,0 +1,9 @@
+﻿using AllOverIt.Collections;
+
+namespace AllOverIt.Serilog.Sinks.CircularBuffer
+{
+    /// <summary>Represents a circular buffer of <see cref="CircularBufferSinkMessage"/> messages.</summary>
+    public interface ICircularBufferSinkMessages : ICircularBuffer<CircularBufferSinkMessage>
+    {
+    }
+}